[main]: finished the  functionality

+++ src/main/java/com/magement/gateway/repositories/device/boundary/DeviceRepository.java
@@ -3,5 +3,5 @@ package com.magement.gateway.repositories.device.boundary;
 import com.magement.gateway.repositories.device.entity.DeviceEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DeviceRepository extends CrudRepository<DeviceEntity, Long> {
+public interface DeviceRepository extends CrudRepository<DeviceEntity, Integer> {
 }
